Harden atomic overwrite recovery and repository hygiene for 1.0.6
