--- .github/workflows/one-shot-105-audit.yml
@@ -1,38 +1,0 @@
-name: Ghost FTP 1.0.5 read-only audit
-
-on:
-  push:
-    branches:
-      - audit/ghost-ftp-1.0.5
-    paths:
-      - .github/workflows/one-shot-105-audit.yml
-
-permissions:
-  contents: read
-
-jobs:
-  audit:
-    runs-on: ubuntu-24.04
-    steps:
-      - name: Checkout source
-        uses: actions/checkout@3d3c42e5aac5ba805825da76410c181273ba90b1
-      - name: Scan exception, write, interface and visible-brand paths
-        shell: bash
-        run: |
-          set -euo pipefail
-          echo '=== GET_MESSAGE ==='
-          grep -RFn --include='*.php' 'getMessage()' 'GhostFTP WEB' || true
-          echo '=== THROWABLE_CATCHES ==='
-          grep -RFn --include='*.php' 'catch (\Throwable' 'GhostFTP WEB' || true
-          echo '=== FILE_PUT_CONTENTS ==='
-          grep -RFn --include='*.php' 'file_put_contents' 'GhostFTP WEB' || true
-          echo '=== REMOTE_WRITE_CALLS ==='
-          grep -RFn --include='*.php' -- '->write(' 'GhostFTP WEB' || true
-          echo '=== REMOTE_WRITE_DECLARATIONS ==='
-          grep -RFn --include='*.php' 'function write(string $remotePath' 'GhostFTP WEB' || true
-          echo '=== RUNTIME_EXCEPTION_DYNAMIC ==='
-          grep -REn --include='*.php' 'new RuntimeException\([^;]*(\$[A-Za-z_][A-Za-z0-9_]*|getMessage\()' 'GhostFTP WEB' || true
-          echo '=== VISIBLE_GHOSTFTP_TOPLEVEL ==='
-          grep -Fn --include='*.php' 'GhostFTP' 'GhostFTP WEB'/*.php || true
-          echo '=== VISIBLE_GHOSTFTP_VIEWS ==='
-          grep -RFn --include='*.php' 'GhostFTP' 'GhostFTP WEB/app/Views' || true
