+++ GhostFTP WEB/VERSION
@@ -1,1 +1,1 @@
-1.0.5
+1.0.6
